+++ src/sp_config_scanner.h
@@ -15,7 +15,7 @@ typedef struct {
   const char *arg; // optional argument / can be not null terminated
   size_t arglen;
   sp_argtype argtype;
-  long lineno;
+  size_t lineno;
 } sp_parsed_keyword;
 
 zend_result sp_config_scan(const char *data, zend_result (*process_rule)(sp_parsed_keyword*));
